+++ src/regex/vm.cpp
@@ -1207,7 +1207,8 @@ bool escape(const StringRef ref, const size_t maxSize, std::string &out) {
     default:
       int codePoint;
       if (unsigned int byteSize = UnicodeUtil::wtf8ToCodePoint(iter, end, codePoint)) {
-        if (ucp::hasPrimeLoneProperty(codePoint, ucp::Lone::ESRegexClassSpace)) {
+        if (ucp::hasPrimeLoneProperty(codePoint, ucp::Lone::ESRegexClassSpace) ||
+            UnicodeUtil::isSurrogate(codePoint)) {
           len = snprintf(buf, std::size(buf), "\\u%04x", codePoint);
         } else {
           memcpy(buf, iter, byteSize);